KVM: X86: Refactor hardware enable/disable operations into a new file
In order to split out the hardware operations into a new module, the first
step is to move them into a new file, which will eventually be part of the
new module. This way there is no code duplication between KVM and the new
module.

This is not a functional change since the code is still built as part of
kvm.o.

Signed-off-by: Anish Ghulati <aghulati@google.com>