KVM: x86: Move user return msr operations out of KVM
Before moving the rest of hardware enable/disable operations into a new
module, the dependencies have to be moved out first. This change moves
user return msr operations into a new file, which will eventually be
converted to a new module.

This is not a functional change.

Updated the code to fix checkpatch warnings.

Signed-off-by: Anish Ghulati <aghulati@google.com>