KVM: X86: Move arch specific hardware operations out of KVM
This is the last of the hardware enable/disable logic for x86 that is
still in KVM. This involves splitting hardware enable/disable out of
kvm-x86-ops.h and into its own header (vac-x86-ops.h) so that the
functions can be called statically.

The code that accounts for unstable TSC is left as a TODO and needs to
be completed.

Signed-off-by: Anish Ghulati <aghulati@google.com>