KVM: VAC: VMX: Remove reference to enable_ept in vac
(TEMPORARY) enable_ept is a module param that governs whether
EPT (two-level paging) is used in VMX or not. It's currently
in KVM, but the hardware enable code needs it to conditionally
sync EPTs. Maybe we'll move it? Or maybe we'll move that phase
of h/w setup.

Fixes one of two build errors; INVEPT_ERROR via the sync is the
other.

Signed-off-by: Venkatesh Srinivas <venkateshs@chromium.org>